Revert "Move DiracProcessManager to avoid test failures"
This reverts commit 0f936aee66557d30ad5352b4d5944a90651d8267.
